### 2019-02-15 #1254 @whedon
[PRE REVIEW]: Phobos: A tool for creating complex robot models
**Submitting author:** @Amudtogal (<a href="http://orcid.org/0000-0002-0190-9780">Simon Reichel</a>)
**Repository:** <a href="https://github.com/dfki-ric/phobos" target ="_blank">https://github.com/dfki-ric/phobos</a>
**Version:** 1.0.0 &quot;Capricious Choutengan&quot;
