### 2019-02-28 #1254 @danielskatz
👋 @arfon - Can you help me see what's wrong with this?  The figure line in [paper.md](https://github.com/dfki-ric/phobos/blob/master/paper.md) looks ok.  Is there a problem with having a link in a caption?  Or the formatting?
