+++ paper.bib
@@ -7,7 +7,7 @@ title = {HyRoDyn: A Modular Software Framework for Solving Analytical Kinematics
 }
 
 @misc{phobosrepo,
-  title = {Phobos Github Repository},
+  title = {{Phobos GitHub Repository}},
   year = {2019},
   howpublished = {\url{https://github.com/dfki-ric/phobos}}
 }
