### 2019-03-01 #1254 @arfon
@danielskatz - looks like we don't support links in captions. That said, anything that looks like a URL us auto-linked in the PDF output anyway.

Fix for the build here: https://github.com/dfki-ric/phobos/pull/178
