### 2019-03-13 #1254 @gkthiruvathukal
@danielskatz Sorry for not seeing your earlier request. I am usually willing, so the answer is yes. It took me a bit of time to get ramped up as an editor but things are going well now.

@kyleniemeyer Yes, I am on the case.

@Amudtogal Do you have any suggested reviewers from https://bit.ly/joss-reviewers? I plan to make an assignment from the list if I do not hear from you within 24 hours.
