### 2019-03-13 #1254 @Amudtogal
In terms of reviewers it would be great to have someone from a simulation & robotics background. The best choice might be **carlojoseRG** with background knowledge in ROS & robotics, however Python is not included in his languages.

Other people having useful knowledge for the review are:
**mosteo** (robotics background, Python)
**trallard** (robotics background, Python)
**hugoledoux** (3D, modelling background, Python)
**CELavecchia** (robotics background, Python)
**traversaro** (simulation & robotics background)
**CameronDevine** (robotics background, Python)
**nnadeau** (robotics & mechanics background, Python)
**jizecn** (robotics background, Python)

I hope this helps making a good choice!
