### 2019-03-16 #1326 @whedon
[REVIEW]: Phobos: A tool for creating complex robot models
**Submitting author:** @Amudtogal (<a href="http://orcid.org/0000-0002-0190-9780">Simon Reichel</a>)
**Repository:** <a href="https://github.com/dfki-ric/phobos" target ="_blank">https://github.com/dfki-ric/phobos</a>
**Version:** v1.0.1
