### 2019-03-23 #1326 @CameronDevine
@Amudtogal Looking through the repository I didn't find any tests or other ways to check if the software is operating correctly (See *Automated tests* section of the checklist above).

I also opened an issue (dfki-ric/phobos#179) about some issues in the Wiki.
