### 2019-03-26 #1326 @Amudtogal
@CameronDevine I will fix the issues in the wiki soon. Sorry for the delay, I had some issues with my computer...

The issue with tests for Phobos has been discussed for a long time. Setting up tests for a Blender addon is a cumbersome process (see https://github.com/dfki-ric/phobos/issues/121). For now, we decided that the additional effort maintaining a CI environment for Phobos is not really worth the trouble. This will be revisited, if we can actually see a growing user base and can also nail down real testing scenarios for the different operators.

Does this make sense, or do you think the tests are absolutely necessary? I'd be ready to discuss this further :)
