### 2019-03-26 #1326 @CameronDevine
@Amudtogal I don't think tests are absolutely necessary. However, I do think it wouldn't be too hard to add in a limited number of tests. For example, `calculateBoxInertia`, `writeURDFGeometry`, and similar functions could be tested without installing blender in a test environment.
