### 2019-03-26 #1326 @gkthiruvathukal
I agree with @CameronDevine. In software engineering, we often refer to these tests as model tests. That is, the part of your code that perhaps isn't entangled with the user interface. If there are patterns being used (e.g. Model-View-Controller) you might be able to focus on the model part. The methods @CameronDevine showed strongly suggest this possibility as the code is independent of Blender per se.
