### 2019-03-27 63682bd
Update sensor definition
