### 2019-04-01 dfki-ric/phobos#180 @CameronDevine
Add note about tilde to the first steps document
It would be worth including in the [first steps](https://github.com/dfki-ric/phobos/wiki/First-Steps) document that `~` cannot be used when setting paths in the Phobos preferences.
