### 2019-04-01 dfki-ric/phobos#181 @CameronDevine
Log file preferences
In the Phobos preferences the "log file path" label is confusing. Originally I though this was the path to a directory where log files should be stored, however I figured out it is the location of the log file itself. I think changing the label to simply "log file" would make this less confusing.
