### 2019-04-02 #1326 @CameronDevine
I opened two more issues today, dfki-ric/phobos#180 and dfki-ric/phobos#181. Both of these issue relate to a minor difficulties I experienced while configuring Phobos.
