### 2019-04-02 dfki-ric/phobos#181 @CameronDevine
A more descriptive error message, such as "Unable to open log file "/.phobos/log/" would also be a reasonable way to fix this issue.
