### 2019-04-16 5701d86
Fix logfile error
The confusing name of the logfile path in the User Preferences is
adjusted.

Also, the log message of a bad logfile location had a bug and did only
show up in the GUI, so it is fixed to print in both GUI and in the
terminal.

Closes #181.
