### 2019-04-16 #1326 @Amudtogal
I have just closed these issues.
Furthermore, I added the model tests to our agenda. Do you expect us to finish the test before advancing the publication?
