### 2019-04-16 dfki-ric/phobos#179 @Amudtogal
I have fixed the merge conflicts, however there are still some TODOs that need to be fixed...
