### 2019-04-21 #1326 @gkthiruvathukal
@CameronDevine, have the issues you reported all been addressed satisfactorily?
@trallard Do you have any additional input on this submission?
