### 2019-04-21 #1326 @CameronDevine
@gkthiruvathukal The only outstanding item for me is the lack of automated tests. Since neither automated or manual tests are provided I do not yet feel comfortable checking that item off.
