### 2019-05-02 #1326 @gkthiruvathukal
@Amudtogal I am seeing that we (or I, more precisely) did not respond to your question about model tests. Sorry for the delay as it has been a busy ending to my academic year here.

Can you please add some basic model tests so we can check off the last box? I'd feel a lot better if there were at least a few tests, which would establish a basis for more expanded testing at a later date. As soon as you can take care of this and @CameronDevine gets a final look, I can proceed.
