### 2019-05-04 #1326 @trallard
HI sorry for taking so long @gkthiruvathukal AND @Amudtogal  from my part the thing missing as well is that corresponding to the tests.

I believe some model tests would be essential even if to have a baseline of the software performance / behaviour at the current state.

Apart from that I do not see any major issues or blockers
