### 2019-05-08 #1326 @gkthiruvathukal
Thanks, @trallard.
@Amudtogal, based on the feedback from the reviewers--and my own relatively strong opinions about the need for tests--please act upon this input so we can proceed to the next steps.
