### 2019-06-04 #1326 @Amudtogal
I am planning to start the tests within the next three weeks. Next weekend there is a big event I am part of, which has been devouring my spare time... When it is over, I can shift my focus back to coding...

Sorry for delaying this whole process (I don't want to waste your time)! Is there a time limit I need to consider or can I pursue this plan?
Thanks again for your patience!
