### 2019-06-14 8811692
Quickfix : Motorconversion
Linked to #185
