### 2019-06-20 9a3b108
Add initial tests & Travis configuration
