+++ tests/testutils/testutils.test.py
@@ -1,0 +1,190 @@
+#!/usr/bin/python3
+
+import phobos
+import unittest
+
+class TestBlenderUtils(unittest.TestCase):
+
+    def test_compileEnumPropertyList(self):
+        testlist = phobos.utils.blender.compileEnumPropertyList([1, 2, 3])
+        self.assertTupleEqual(testlist, ((1, 1, 1), (2, 2, 2), (3, 3, 3)))
+
+class TestGeneralUtils(unittest.TestCase):
+
+    def test_is_float(self):
+        self.assertTrue(phobos.utils.general.is_float('12.3'))
+        self.assertFalse(phobos.utils.general.is_float('hello'))
+        self.assertFalse(phobos.utils.general.is_float('[12.3, 3.2]'))
+
+    def test_is_int(self):
+        self.assertTrue(phobos.utils.general.is_int('12'))
+        self.assertFalse(phobos.utils.general.is_int('hello'))
+        self.assertFalse(phobos.utils.general.is_int('12.3452234'))
+        self.assertFalse(phobos.utils.general.is_int('[12, 3]'))
+
+    def test_parse_number(self):
+        number = phobos.utils.general.parse_number('12')
+        self.assertEqual(type(number), type(1))
+        number = phobos.utils.general.parse_number('12.124')
+        self.assertEqual(type(number), type(12.124))
+        number = phobos.utils.general.parse_number('hello')
+        self.assertEqual(type(number), type('hello'))
+
+    def test_only_contains_int(self):
+        testlist = [1, 2, 3, 4, 5]
+        self.assertTrue(phobos.utils.general.only_contains_int(testlist))
+        testlist = [1, 2, 3, 4, 5, 12.3]
+        self.assertFalse(phobos.utils.general.only_contains_int(testlist))
+        testlist = [1, 2, 3, 4, 5, 'hello']
+        self.assertFalse(phobos.utils.general.only_contains_int(testlist))
+
+    def test_only_contains_float(self):
+        testlist = [1.1, 2.1, 3.5, 4.7, 5.6]
+        self.assertTrue(phobos.utils.general.only_contains_float(testlist))
+        testlist = [1.1, 2.1, 3.5, 4.7, 5.6, 12]
+        self.assertFalse(phobos.utils.general.only_contains_float(testlist))
+        testlist = [1.1, 2.1, 3.5, 4.7, 5.6, 'hello']
+        self.assertFalse(phobos.utils.general.only_contains_float(testlist))
+
+    def test_parse_text(self):
+        teststring = '1 2 3 4 5'
+        self.assertListEqual(phobos.utils.general.parse_text(teststring), [1, 2, 3, 4, 5])
+        teststring = '1.1 2.1 3.4 4.7 5.1'
+        self.assertListEqual(phobos.utils.general.parse_text(teststring), [1.1, 2.1, 3.4, 4.7, 5.1])
+        teststring = 'a bc de hello whatsoever'
+        self.assertListEqual(phobos.utils.general.parse_text(teststring), ['a', 'bc', 'de', 'hello', 'whatsoever'])
+        teststring = '12.3'
+        self.assertEqual(phobos.utils.general.parse_text(teststring), 12.3)
+        teststring = '12'
+        self.assertEqual(phobos.utils.general.parse_text(teststring), 12)
+
+    def test_calcBoundingBoxCenter(self):
+        corners = [
+            [0., 0., 0.],
+            [1., 0., 0.],
+            [1., 0., 1.],
+            [0., 0., 1.],
+            [0., 1., 0.],
+            [1., 1., 0.],
+            [1., 1., 1.],
+            [0., 1., 1.]
+        ]
[…]
