+++ .travis.yml
@@ -1,0 +1,49 @@
+# this is essential to make python work out of the box
+language: python
+
+# safelist
+branches:
+  only:
+  - testingsuite
+
+# Define which python version to build against.
+# As blender comes bundled with it's own python you can use just one version if you like
+python:
+  - "2.7"
+  - "3.5"
+
+before_install:
+  # update apt-gets repository sources
+  - sudo apt-get update -qq
+
+  # install blender from official sources.
+  # This will most propably install an outdated blender version,
+  # but it will resolve all system dependencies blender has to be able to run.
+  - sudo apt-get install blender
+
+install:
+  # create temp directory where we store a recent blender version
+  - mkdir tmp && cd tmp
+
+  # download the blender version you want to test against
+  - wget http://mirror.cs.umn.edu/blender.org/release/Blender2.79/blender-2.79-linux-glibc219-x86_64.tar.bz2
+
+  # Extract the archive
+  - tar jxf blender-2.79-linux-glibc219-x86_64.tar.bz2
+
+  # rename the extracted folder to "blender" for easier reference
+  - mv blender-2.79-linux-glibc219-x86_64.tar.bz2 blender
+
+  # remove the archive, we do not need it anymore
+  - rm blender-2.79-linux-glibc219-x86_64.tar.bz2
+
+  # go back to root directory
+  - cd ..
+
+  # now create a symlink to the addon within blenders addons directory
+  # this is important, because otherwhise blender would not be able to load the addon
+  - sudo ln -s ${PWD}/phobos ${PWD}/tmp/blender/2.74/scripts/addons/phobos
+
+# Finally start our test runner passing it the blender executable from the downloaded blender release
+script: python testrunner.py ./tmp/blender/blender
+
