### 2019-06-25 00d5ad0
Add gitignore exception for travis config
