+++ .travis.yml
@@ -4,7 +4,7 @@ language: python
 # safelist
 branches:
   only:
-  - testingsuite
+  - master
 
 # Define which python version to build against.
 # As blender comes bundled with it's own python you can use just one version if you like
@@ -32,7 +32,7 @@ install:
   - tar jxf blender-2.79-linux-glibc219-x86_64.tar.bz2
 
   # rename the extracted folder to "blender" for easier reference
-  - mv blender-2.79-linux-glibc219-x86_64.tar.bz2 blender
+  - mv blender-2.79-linux-glibc219-x86_64 blender
 
   # remove the archive, we do not need it anymore
   - rm blender-2.79-linux-glibc219-x86_64.tar.bz2
@@ -42,7 +42,7 @@ install:
 
   # now create a symlink to the addon within blenders addons directory
   # this is important, because otherwhise blender would not be able to load the addon
-  - sudo ln -s ${PWD}/phobos ${PWD}/tmp/blender/2.74/scripts/addons/phobos
+  - sudo ln -s ${PWD}/phobos ${PWD}/tmp/blender/2.79/scripts/addons/phobos
 
 # Finally start our test runner passing it the blender executable from the downloaded blender release
 script: python testrunner.py ./tmp/blender/blender
