### 2019-06-25 ac4847b
Add installto parameter to setup script and apply to travis installation
