+++ .travis.yml
@@ -9,7 +9,6 @@ branches:
 # Define which python version to build against.
 # As blender comes bundled with it's own python you can use just one version if you like
 python:
-  - "2.7"
   - "3.5"
 
 before_install:
@@ -42,7 +41,7 @@ install:
 
   # now create a symlink to the addon within blenders addons directory
   # this is important, because otherwhise blender would not be able to load the addon
-  - sudo ln -s ${PWD}/phobos ${PWD}/tmp/blender/2.79/scripts/addons/phobos
+  - sudo python3.5 ${PWD}/setup.py --installto ${PWD}/tmp/blender/2.79/scripts/addons/phobos
 
 # Finally start our test runner passing it the blender executable from the downloaded blender release
 script: python testrunner.py ./tmp/blender/blender
