### 2019-06-25 21f6300
Add installation prereqs and change to xenial
