### 2019-06-25 3ed4239
Add Travis configuration
