+++ .travis.yml
@@ -11,7 +11,7 @@ branches:
 # Define which python version to build against.
 # As blender comes bundled with it's own python you can use just one version if you like
 python:
-  - "3.5"
+  - "3.5.7"
 
 before_install:
   # update apt-gets repository sources
@@ -24,9 +24,9 @@ before_install:
 
   - sudo apt-get install python-pip
 
-  - sudo pip install PyYaml
+  - pip install PyYaml
 
-  - sudo pip install numpy
+  - pip install numpy
 
 install:
   # create temp directory where we store a recent blender version
@@ -49,7 +49,11 @@ install:
 
   # now create a symlink to the addon within blenders addons directory
   # this is important, because otherwhise blender would not be able to load the addon
-  - sudo python ${PWD}/setup.py --installto ${PWD}/tmp/blender/2.79/scripts/addons/phobos
+  # - sudo ln -s ${PWD}/phobos ${PWD}/tmp/blender/2.79/scripts/addons/phobos
+
+  # instead we use the setup.py to install the addon to the folder
+  - python ${PWD}/setup.py --installto ${PWD}/tmp/blender/2.79/scripts/addons/phobos
+  - ls ${PWD}/tmp/blender/2.79/scripts/addons/phobos
 
 # Finally start our test runner passing it the blender executable from the downloaded blender release
 script: python testrunner.py ./tmp/blender/blender
