### 2019-06-26 c8092d7
Fix site package bug in setup script
As *site* is broken in virtual environments, the setup script would fail
in these circumstances. We use the yaml path instead to identify the
proper path.
