### 2019-07-01 #1326 @Amudtogal
Hi folks,
a short update on my progress:

I have setup a Travis CI, which runs and installs Phobos and is able to run different tests from the test folder.
https://travis-ci.org/Amudtogal/phobos
Right now, all these changes sit on my fork on github (https://github.com/Amudtogal/phobos), as there are still some minor issues to be resolved before I can merge it into the master:
- proper error code propagation from the test files to the Travis commandline (otherwise the build never fails)
- fix the current tests (I wrote a bunch of tests which do not rely on any of the Blender functionality)

Let me know whether you think this is promising or you expected something else :)
Have a nice week,
Simon
