### 2019-07-01 #1326 @gkthiruvathukal
@trallard Thanks for checking in! We definitely can use your help with the remaining work on this review. Would you be willing to take a look at @Amudtogal's updates for Travis CI/testing? If it is on the right track, I think we can move toward acceptance pretty soon.
