### 2019-07-01 #1326 @CameronDevine
@Amudtogal, once the error code propagation is fixed and tests for model joints are added, as mentioned in `testmodel.test.py`, I will be ready to recommend this for publication.
