+++ testrunner.py
@@ -14,4 +14,8 @@ if len(sys.argv) > 1:
 # iterate over each *.test.blend file in the "tests" directory
 # and open up blender with the .test.blend file and the corresponding .test.py python script
 for file in glob.glob('tests/**/*.test.blend'):
-    subprocess.call([blenderExecutable, '--addons', 'phobos', '--factory-startup', '-noaudio', '-b', file, '--python', file.replace('.blend', '.py')])
+    code = subprocess.call([blenderExecutable, '--addons', 'phobos', '--factory-startup', '-noaudio', '-b', file, '--python', file.replace('.blend', '.py'), '--python-exit-code', '1'])
+    if code:
+        sys.exit(code)
+
+sys.exit(0)
