+++ tests/testutils/testutils.test.py
@@ -1,190 +1,195 @@
 #!/usr/bin/python3
 
-import phobos
-import unittest
+import sys
 
-class TestBlenderUtils(unittest.TestCase):
+try:
+    import phobos
+    import unittest
 
-    def test_compileEnumPropertyList(self):
-        testlist = phobos.utils.blender.compileEnumPropertyList([1, 2, 3])
-        self.assertTupleEqual(testlist, ((1, 1, 1), (2, 2, 2), (3, 3, 3)))
+    class TestBlenderUtils(unittest.TestCase):
 
-class TestGeneralUtils(unittest.TestCase):
+        def test_compileEnumPropertyList(self):
+            testlist = phobos.utils.blender.compileEnumPropertyList([1, 2, 3])
+            self.assertTupleEqual(testlist, ((1, 1, 1), (2, 2, 2), (3, 3, 3)))
 
-    def test_is_float(self):
-        self.assertTrue(phobos.utils.general.is_float('12.3'))
-        self.assertFalse(phobos.utils.general.is_float('hello'))
-        self.assertFalse(phobos.utils.general.is_float('[12.3, 3.2]'))
+    class TestGeneralUtils(unittest.TestCase):
 
-    def test_is_int(self):
-        self.assertTrue(phobos.utils.general.is_int('12'))
-        self.assertFalse(phobos.utils.general.is_int('hello'))
-        self.assertFalse(phobos.utils.general.is_int('12.3452234'))
-        self.assertFalse(phobos.utils.general.is_int('[12, 3]'))
+        def test_is_float(self):
+            self.assertTrue(phobos.utils.general.is_float('12.3'))
+            self.assertFalse(phobos.utils.general.is_float('hello'))
+            self.assertFalse(phobos.utils.general.is_float('[12.3, 3.2]'))
 
-    def test_parse_number(self):
-        number = phobos.utils.general.parse_number('12')
-        self.assertEqual(type(number), type(1))
-        number = phobos.utils.general.parse_number('12.124')
-        self.assertEqual(type(number), type(12.124))
-        number = phobos.utils.general.parse_number('hello')
-        self.assertEqual(type(number), type('hello'))
+        def test_is_int(self):
+            self.assertTrue(phobos.utils.general.is_int('12'))
+            self.assertFalse(phobos.utils.general.is_int('hello'))
+            self.assertFalse(phobos.utils.general.is_int('12.3452234'))
+            self.assertFalse(phobos.utils.general.is_int('[12, 3]'))
 
-    def test_only_contains_int(self):
-        testlist = [1, 2, 3, 4, 5]
-        self.assertTrue(phobos.utils.general.only_contains_int(testlist))
-        testlist = [1, 2, 3, 4, 5, 12.3]
-        self.assertFalse(phobos.utils.general.only_contains_int(testlist))
-        testlist = [1, 2, 3, 4, 5, 'hello']
-        self.assertFalse(phobos.utils.general.only_contains_int(testlist))
+        def test_parse_number(self):
+            number = phobos.utils.general.parse_number('12')
+            self.assertEqual(type(number), type(1))
+            number = phobos.utils.general.parse_number('12.124')
+            self.assertEqual(type(number), type(12.124))
+            number = phobos.utils.general.parse_number('hello')
+            self.assertEqual(type(number), type('hello'))
 
-    def test_only_contains_float(self):
-        testlist = [1.1, 2.1, 3.5, 4.7, 5.6]
-        self.assertTrue(phobos.utils.general.only_contains_float(testlist))
[…]
