### 2019-07-15 49bf625
Add exception catch for compiling mechanism/model libraries
