### 2019-07-15 15e9820
Fix test error/failure propagation
