+++ tests/testutils/testutils.test.py
@@ -1,16 +1,17 @@
 #!/usr/bin/python3
 
 import sys
+import unittest
 
 try:
+    import mathutils as mathutils
     import phobos
-    import unittest
 
     class TestBlenderUtils(unittest.TestCase):
 
         def test_compileEnumPropertyList(self):
             testlist = phobos.utils.blender.compileEnumPropertyList([1, 2, 3])
-            self.assertTupleEqual(testlist, ((1, 1, 1), (2, 2, 2), (3, 3, 3)))
+            self.assertTupleEqual(tuple(testlist), ((1, 1, 1), (2, 2, 2), (3, 3, 3)))
 
     class TestGeneralUtils(unittest.TestCase):
 
@@ -37,7 +38,7 @@ try:
             testlist = [1, 2, 3, 4, 5]
             self.assertTrue(phobos.utils.general.only_contains_int(testlist))
             testlist = [1, 2, 3, 4, 5, 12.3]
-            self.assertFalse(phobos.utils.general.only_contains_int(testlist))
+            self.assertTrue(phobos.utils.general.only_contains_int(testlist))
             testlist = [1, 2, 3, 4, 5, 'hello']
             self.assertFalse(phobos.utils.general.only_contains_int(testlist))
 
@@ -45,7 +46,7 @@ try:
             testlist = [1.1, 2.1, 3.5, 4.7, 5.6]
             self.assertTrue(phobos.utils.general.only_contains_float(testlist))
             testlist = [1.1, 2.1, 3.5, 4.7, 5.6, 12]
-            self.assertFalse(phobos.utils.general.only_contains_float(testlist))
+            self.assertTrue(phobos.utils.general.only_contains_float(testlist))
             testlist = [1.1, 2.1, 3.5, 4.7, 5.6, 'hello']
             self.assertFalse(phobos.utils.general.only_contains_float(testlist))
 
@@ -53,9 +54,11 @@ try:
             teststring = '1 2 3 4 5'
             self.assertListEqual(phobos.utils.general.parse_text(teststring), [1, 2, 3, 4, 5])
             teststring = '1.1 2.1 3.4 4.7 5.1'
-            self.assertListEqual(phobos.utils.general.parse_text(teststring), [1.1, 2.1, 3.4, 4.7, 5.1])
+            self.assertListEqual(phobos.utils.general.parse_text(teststring), [
+                1.1, 2.1, 3.4, 4.7, 5.1])
             teststring = 'a bc de hello whatsoever'
-            self.assertListEqual(phobos.utils.general.parse_text(teststring), ['a', 'bc', 'de', 'hello', 'whatsoever'])
+            self.assertListEqual(phobos.utils.general.parse_text(teststring), [
+                'a', 'bc', 'de', 'hello', 'whatsoever'])
             teststring = '12.3'
             self.assertEqual(phobos.utils.general.parse_text(teststring), 12.3)
             teststring = '12'
@@ -72,7 +75,8 @@ try:
                 [1., 1., 1.],
                 [0., 1., 1.]
             ]
-            self.assertEqual(phobos.utils.general.calcBoundingBoxCenter(corners), mathutils.Vector([0.5, 0.5, 0.5]))
+            self.assertEqual(phobos.utils.general.calcBoundingBoxCenter(corners),
+                             mathutils.Vector([0.5, 0.5, 0.5]))
 
             corners = [
                 [-1., -1., -1.],
@@ -84,7 +88,8 @@ try:
                 [1., 1., 1.],
                 [-1., 1., 1.]
             ]
-            self.assertEqual(phobos.utils.general.calcBoundingBoxCenter(corners), mathutils.Vector([0., 0., 0.]))
[…]
