### 2019-07-18 #1326 @Amudtogal
@gkthiruvathukal Sorry, I didn't propagate the information over here. I did fix the error propagation based on the [Pull Request from CameronDevine](https://github.com/Amudtogal/phobos/pull/1). Next up will be fixing the errors within the test scripts. After that, I will apply some polishing to the messy git history on my fork and push it to the Phobos master.
The current build (finally failing when tests fail) can be found here:
https://travis-ci.org/Amudtogal/phobos/builds/559186686
