### 2019-07-24 798256c
Fixed mimic motors for SMURF format
See #189 for details
