### 2019-07-24 75b06f4
Fixed inertia validiation
Decimal places where wrong
