### 2019-07-24 5e31c5e
Updated camera sensor definition of hud_pos to hud_idx.
