### 2019-07-27 19b6c3a
Allow negative index for l2str util function
