### 2019-08-03 09cf493
Add zero string length criterion to IsValidModelname
