### 2019-08-03 #1326 @Amudtogal
As of now, the tests are on the master branch :smiley:
It might take a week until the Travis Setup is activated (as I do not have access to the phobos-owning github account), however the e-mail for this is sent already.
The latest build can be found on my fork Travis (https://travis-ci.org/Amudtogal/phobos) and works as intended.
As soon as the Travis Setup is activated, I will write again!
Thanks for the patience, we are close to finishing :+1:
