### 2019-08-03 #1326 @gkthiruvathukal
@Amudtogal Ok, great! I am already happy with what you're demonstrating here (your latest build). Hopefully, you can get the photos owner to authorize Travis. Then approval will be almost spontaneous. I'm looking forward to accepting your submission and appreciate your work on the last mile (or last kilometer). We're almost there.
