### 2019-08-12 #1326 @Amudtogal
We finally reached the end of the road!
The first build on Phobos ran successfully: https://travis-ci.org/dfki-ric/phobos
Enjoy your break!
I am looking forward to your reply when you return :)
