### 2019-08-12 #1326 @CameronDevine
@gkthiruvathukal All of the items on my checklist are now complete. I believe this paper is ready to be accepted.
