### 2019-08-25 #1326 @gkthiruvathukal
Ok, I'm back. Now I just need to get the semester started.

Thanks @CameronDevine for confirming that all items on your checklist are now complete.

@Amudtogal, I am ready to move toward final acceptance. Please do the following:
