### 2019-08-28 #1326 @Amudtogal
Great!
The last steps will however take another two weeks... I am sorry for another delay.
This time it is my own thesis getting in the way :smile:
I expect we will draft the next release mid september, including the tests and changes that have been added since version 1.0.0. As soon as we are finished with it, I will come back to you!
Have a great start for the semester, and I am looking forward to the last few steps!
