### 2019-10-03 #1326 @Amudtogal
We are running through the last stages, I will keep you updated!
There will be a change from LGPL to BSD, but I guess that is no problem...
