### 2019-10-15 #1326 @gkthiruvathukal
@Amudtogal Can you let us know when this license change is complete? Once done, I believe we are ready to move toward acceptance. You'll need to tag the latest release and create the Zenodo artifact.
