### 2019-10-30 #1326 @danielskatz
👋 @Amudtogal - I'm just checking in on this as the associate-editor-in-chief on duty this week ... Is there any progress on the license change you can report?
