### 2019-12-16 a1fb734
Change license to BSD3
Change file licenses, cleanup shebangs
