### 2019-12-16 f12071b
Add version target for help with versioning to Makefile
