+++ Makefile
@@ -8,16 +8,17 @@
 
 help:
 		@echo	'Targets of Phobos:'
-		@echo 	'  help      - Print this help information.'
-		@echo 	'  init      - Install the Phobos requirements using pip.'
-		@echo	'  install   - Install the Phobos code to your Blender installation.'
-		@echo	'              This also sets up the configuration folder for Phobos.'
+		@echo 	'  apidoc    - Generates the Sphinx API documentation and moves'
+		@echo	'              it to the gh-pages branch.'
 		@echo	'  clean     - Removes the installation configuration file for Phobos.'
 		@echo	'              This does not remove the Phobos installation and configurations!'
 		@echo 	'  format    - Formats the python code in the folder using the black code'
 		@echo 	'              formatter (github.com/ambv/black).'
-		@echo 	'  apidoc    - Generates the Sphinx API documentation and moves it to the gh-pages'
-		@echo 	'              branch.'
+		@echo 	'  help      - Print this help information.'
+		@echo 	'  init      - Install the Phobos requirements using pip.'
+		@echo	'  install   - Install the Phobos code to your Blender installation.'
+		@echo	'              This also sets up the configuration folder for Phobos.'
+		@echo	'  version   - Prints some help relating to drafting a new version.'
 
 init:
 		pip install -r requirements.txt
@@ -46,5 +47,13 @@ apidoc:
 		rm -rf *.html *.js *.inv _sources _static
 		mv doc/_build/html/* .
 		@echo 'Please commit and push the changes to publish the new doc on https://dfki-ric.github.io/phobos'
+
+version:
+		@echo 'Change version in:'
+		@echo '  - codemeta.json'
+		@echo '  - phobos/defs.py'
+		@echo '  - phobos/__init__.py'
+		@echo '  - doc/conf.py'
+		@echo '  - doc/index.rst'
 
 .PHONY: init test install format apidoc help
