### 2019-12-16 f6d3304
Update versions in defs & codemeta
