+++ phobos/__init__.py
@@ -71,8 +71,8 @@ bl_info = {
     "name": "Phobos",
     "description": "A toolbox to enable editing of robot models in Blender.",
     "author": "Kai von Szadkowski, Ole Schwiegert, Stefan Rahms, Malte Langosz, Simon Reichel",
-    "version": (0, 7, 1),
-    "blender": (2, 69, 0),
+    "version": (1, 0, 1),
+    "blender": (2, 79, 0),
     "location": "Phobos adds a number of custom tool panels.",
     "warning": "",
     "wiki_url": "https://github.com/dfki-ric/phobos/wiki",
