+++ doc/conf.py
@@ -71,7 +71,7 @@ copyright = u'2018, DFKI GmbH'
 # The short X.Y version.
 version = '1.0'
 # The full version, including alpha/beta/rc tags.
-release = '1.0.0  "Capricious Choutengan"'
+release = '1.0.1  "Capricious Choutengan"'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
