### 2019-12-16 bd6c297
Update master version in defs & codemeta
