### 2019-12-16 a045256
Fix bug in phobos model parsing
