### 2019-12-16 9fafbc3
Cleanup and improve installation by using Blender's python
This is related to #200.

The python_dist_packages.conf was replaced by installation.conf, which
contains much more information, such as the Blender installation path,
the Blender Python version etc.

The required Python packages are installed directly through Blender's
Python, which obsoletes the manual installation.

At the same time, we do not need to add the Python path to Blender
anymore, as we can install dependencies to the Blender path, which
allows us to maintain a curated package list independently from the
system.

To make i
[…]
