+++ setup.py
@@ -19,36 +19,21 @@ Make sure you run it with the appropriate python version (3.5).
 
 import os
 import os.path as path
+import glob
 import sys
 import shutil
+import argparse
 from distutils.dir_util import copy_tree
-import importlib
-
-scriptinformation = """
-This is the setup script for Phobos.
-
-By default, it installs the Blender AddOn Phobos to the Blender configuration folder.
-At the same time, the default configurations are copied to the user Phobos configuration
-folder.
-
-Parameters:
-
-    --help: Show this message and exit.
-
-    --installto DIR: install to a different folder
-
-    --startup-preset: Copies the default Phobos Blender startup file to the Blender
-        configuration folder (replacing the existing startup file).
-"""
+from importlib import util
 
 # make installation originate from the path of this setup file
-phoboshome = os.path.dirname(os.path.abspath(__file__))
+phoboshome = path.dirname(path.abspath(__file__))
 
 # load the phobossystem as module from file
-module_spec = importlib.util.spec_from_file_location(
+module_spec = util.spec_from_file_location(
     'phobossystem', path.join(phoboshome, 'phobos/phobossystem.py')
 )
-phobossystem = importlib.util.module_from_spec(module_spec)
+phobossystem = util.module_from_spec(module_spec)
 module_spec.loader.exec_module(phobossystem)
 addonpath = path.join(phobossystem.getScriptsPath(), 'addons', 'phobos')
 blenderconfigpath = phobossystem.getBlenderConfigPath()
@@ -68,52 +53,117 @@ def updateFolderContents(src, dst):
 
 
 if __name__ == '__main__':
-    # check whether the right python version is used
-    pyver = sys.version_info
-    print(
-        'You started this script with python version: {}.{}'.format(
-            str(pyver.major), str(pyver.minor)
-        )
-    )
-    print('Make sure Blender uses the same version!')
 
-    if '--help' in sys.argv:
-        print(scriptinformation)
-        sys.exit(0)
+    parser = argparse.ArgumentParser(description='''
+        This is the setup script for Phobos.
 
-    if '--installto' in sys.argv:
-        index = sys.argv.index('--installto')
+        By default, it installs the Blender AddOn Phobos to the Blender configuration folder.
+        At the same time, the default configurations are copied to the user Phobos configuration
+        folder.''')
+    parser.add_argument('-p', '--startup-preset', dest='startup_preset',
+                        action='store_true', default=False,
+                        help='''Copies the default Phobos Blender startup file to the Blender
+                            'configuration folder (replacing the existing startup file).''')
+    parser.add_argument('-t', '--install-to', dest='install_to', metavar='DIR', default=None,
+                        help='install addon to the specified directory')
+    parser.add_argument('-b', '--blender', metavar='DIR', dest='blender_path',
+                        default=None, help='use the specified Blender folder (instead of asking)')
 
-        addonpath = sys.argv[index + 1]
+    globals().update(vars(parser.parse_args()))
+
+    # look for existing installation configuration
+    if path.isfile(path.join(phoboshome, 'installation.conf')):
+        print('installation.conf found! Configuration done.')
+        with open(path.join(phoboshome, 'installation.conf'), 'r') as conffile:
+            python_package_path = conffile.readline().split(' #')[0]
+            if not blender_path:
+                blender_path = conffile.readline().split(' #')[0]
+            else:
+                conffile.readline()
+            python_executable = conffile.readline().split(' #')[0]
+            blender_executable = conffile.readline().split(' #')[0]
+            python_version = conffile.readline().split(' #')[0]
+            blender_version = conffile.readline().split(' #')[0]
+    # check for existing YAML installation
+    else:
+        if not blender_path:
+            blender_path = path.expanduser(input('Where is Blender installed to? '))
[…]
