+++ Makefile
@@ -27,7 +27,7 @@ install:
 		python3 setup.py
 
 clean:
-		rm python_dist_packages.conf
+		rm installation.conf
 
 format:
 		# add docstrings and format to google style
