### 2019-12-16 22e20d4
Add debug info option to setup.py
By using `--info`, the current installation configuration will be
printed, which can be used to debug installation issues in the future.
