### 2019-12-17 500c763
Fix travis build installation call
