### 2019-12-17 07ab803
Fix installation path of addon for -t option
