### 2019-12-17 faab42d
Fix copying of installation.conf
