### 2019-12-17 f160135
Add orcid to paper and remove Julius
