### 2019-12-17 #1326 @Amudtogal
We reached the final stage:
 **Where can I do this? The release tag is 1.0.1**
- [ ] List the Zenodo DOI of the archived version here. **10.5281/zenodo.3581101**

I haven't published the Zenodo archive yet, as there are a few options I could fill in (e.g. the Journal section). Just give me a call what I shall do with it and I can get it online!
