### 2020-01-07 #1326 @gkthiruvathukal
@Amudtogal Happy new year & thanks for the gentle nudge. Just getting back into the swing of things after the holiday/new year (and getting ready for the upcoming academic semester).

@openjournals/joss-editors I am ready to proceed toward acceptance of this submission.

Thanks to @CameronDevine and @trallard for your reviews!
