### 2020-01-07 #1326 @terrytangyuan
@gkthiruvathukal I think you want to use `@openjournals/joss-editors` instead of `@openjournals/joss-editors`.
