### 2020-01-07 #1326 @gkthiruvathukal
@openjournals/joss-eics  I am ready to proceed toward acceptance of this submission.  My bad for not checking the completion properly (editors instead of eics).
