### 2020-01-07 #1326 @danielskatz
👋 @arfon - note that ref 2 in the xml has the same problem as we've seen before with an incomplete URL...
This is otherwise ready for accepting.
