### 2020-01-07 #1326 @kthyng
@danielskatz I'm worried that I am missing problems in the XML when finishing up submissions. What do you see that is incorrect?
