### 2020-01-07 #1326 @danielskatz
see line 89 of the XML in https://github.com/openjournals/joss-papers/pull/1204/files where the "https" is missing at the start of the URL
