### 2020-01-07 #1326 @Amudtogal
Thanks everyone!
Just checking back:
The Zenodo Archive is currently a *draft*, I have not published it, because I was not sure, whether I should put any information about the Journal into the Zenodo form...

> We reached the final stage:
>
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
>   **Where can I do this? The release tag is 1.0.1 "Capricious Choutengan"**
> * [x]  Archive the reviewed software in Zenodo
> * [x]  Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
> * [x]  List the Zenodo DOI of the archived version here. **10.5281/zenodo.3581101**
>
> I haven't published the Zenodo archive yet, as there are a few options I could fill in (e.g. the Journal section). Just give me a call what I shall do with it and I can get it online!

Shall I go online or fill in this stuff:
![joss](https://user-images.githubusercontent.com/18306764/71910387-bbc81300-3169-11ea-9215-71d9dbdbf956.png)
