### 2020-01-07 #1326 @arfon
> 👋 @arfon - can you fix the XML, and then close this?

This is fixed.

@Amudtogal - I don't think you need to put the journal information into this form.
